server/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
groq
kokoro-onnx
soundfile
numpy
torch
scipy
